requirement.txt
flask
python-dotenv
python-jose
flask-cors
six
Authlib
auth0-python
Flask-Session